Refactor demo
