+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
 
     <variable
       name="viewModel"
-      type="com.skydoves.sandwichdemo.MainViewModel" />
+      type="com.skydoves.sandwichdemo.datasource.MainDataSourceViewModel" />
 
     <variable
       name="adapter"
