+++ app/src/main/res/layout/activity_main_coroutines_operator.xml
@@ -7,7 +7,7 @@
 
     <variable
       name="viewModel"
-      type="com.skydoves.sandwichdemo.operator.MainCoroutinesOperatorViewModel" />
+      type="com.skydoves.sandwichdemo.operator.MainOperatorViewModel" />
 
     <variable
       name="adapter"
